requirements.txt
appdirs==1.4.3
packaging==16.8
pyparsing==2.2.0
python-chess==0.18.1
six==1.10.0
